// src/app/shared/models/jobDetail.ts
export class JobDetailModel{
    ESPJobName: string = '';
    SQLJobName: string = '';
    applicationName: string = '';
    shortDescription: string = '';
    timeOut: string = '';
    upstreamJob: string = '';
    frequency: string = '';
    executionTime: string = '';
    downstreamJob: string = '';
}